BENTO-2279 Refactor Global Search search results
