--- src/bento-core/GlobalSearch/SearchPage/components/cards/progamCard.js
@@ -1,92 +1,0 @@
-import { Grid, withStyles } from '@material-ui/core';
-import React from 'react';
-import { Link } from 'react-router-dom';
-import { prepareLinks } from 'bento-components';
-import PropertyItem from '../propertyItem';
-
-const ProgamCard = ({ data, classes, index }) => {
-  const properties = [
-
-    {
-      label: 'Program Code',
-      dataField: 'program_code',
-
-    },
-
-    {
-      label: 'Program Name',
-      dataField: 'program_name',
-    },
-
-  ];
-
-  const propertiesWithLinks = prepareLinks(properties, data);
-
-  return (
-    <>
-      <Grid item container className={classes.card}>
-        <Grid item xs={1} className={classes.indexContainer}>
-          {index + 1 }
-        </Grid>
-        <Grid item xs={11} className={classes.propertyContainer}>
-          <div>
-            <span className={classes.detailContainerHeader}>Program</span>
-            <Link to={`/program/${data.program_id}`} className={classes.cardTitle}>
-              {data.program_id}
-            </Link>
-
-          </div>
-
-          {propertiesWithLinks.map((prop) => (
-            <PropertyItem
-              label={prop.label}
-              value={data[prop.dataField]}
-              link={prop.link}
-            // labelLink={prop.labelLink}
-            // classes={classes}
-              index
-            />
-          ))}
-        </Grid>
-      </Grid>
-    </>
-  );
-};
-
-const styles = (theme) => ({
-  indexContainer: {
-    padding: '18px 0px 18px 18px',
-    color: '#747474',
-    fontFamily: 'Inter',
-    fontSize: '13px',
-  },
-  propertyContainer: {
-    padding: '16px 16px 16px 0px',
-    borderBottom: '2px solid #E7EEF5',
-  },
-  cardTitle: {
-    color: theme.palette.text.link,
-    textDecoration: 'none',
-    fontSize: '16px',
-    fontFamily: 'Nunito',
-    paddingLeft: '9px',
-    verticalAlign: 'middle',
-  },
-  content: {
-    fontSize: '12px',
-  },
-  detailContainerHeader: {
-    textTransform: 'uppercase',
-    padding: '2px 8px',
-    backgroundColor: '#FFE25A',
-    color: '#000000',
-    fontFamily: 'Nunito',
-    fontSize: '12px',
-    fontWeight: '400',
-    letterSpacing: '0.9px',
-    verticalAlign: 'middle',
-    borderRadius: '4px',
-  },
-});
-
-export default withStyles(styles, { withTheme: true })(ProgamCard);
