+++ src/bento-core/GlobalSearch/SearchResults/components/Cards/ValueCard.js
@@ -1,0 +1,98 @@
+import { Grid, withStyles } from '@material-ui/core';
+import React from 'react';
+import { prepareLinks } from 'bento-components';
+import PropertyItem from './PropertyItem';
+
+const CARD_PROPERTIES = [
+  {
+    label: 'Data Model Node',
+    dataField: 'node_name',
+  },
+  {
+    label: 'Property Name',
+    dataField: 'property_name',
+  },
+  {
+    label: 'Property Description',
+    dataField: 'property_description',
+  },
+  {
+    label: 'Property Required',
+    dataField: 'property_required',
+  },
+  {
+    label: 'Property Type',
+    dataField: 'property_description',
+  },
+  {
+    label: 'Property Type',
+    dataField: 'value',
+  },
+  {
+    label: 'Page Link',
+    dataField: 'node_name',
+    link: '/resources',
+  },
+];
+
+const ValueCard = ({ data, classes, index }) => {
+  const propertiesWithLinks = prepareLinks(CARD_PROPERTIES, data);
+
+  return (
+    <Grid item container className={classes.card}>
+      <Grid item xs={1} className={classes.indexContainer}>
+        {index + 1 }
+      </Grid>
+      <Grid item xs={11} className={classes.propertyContainer}>
+        <div>
+          <span className={classes.detailContainerHeader}>DATA MODEL</span>
+          <span className={classes.cardTitle}>
+            {data.highlight}
+          </span>
+        </div>
+        {propertiesWithLinks.map((prop, idx) => (
+          <PropertyItem
+            index={idx}
+            label={prop.label}
+            value={data[prop.dataField]}
+            link={prop.link}
+          />
+        ))}
+      </Grid>
+    </Grid>
+  );
+};
+
+const styles = (theme) => ({
+  indexContainer: {
+    padding: '18px 0px 18px 18px',
+    color: '#747474',
+    fontFamily: 'Inter',
+    fontSize: '13px',
+  },
+  propertyContainer: {
+    padding: '16px 16px 16px 0px',
+    borderBottom: '2px solid #E7EEF5',
+  },
+  cardTitle: {
+    color: theme.palette.text.link,
+    fontSize: '16px',
+    fontFamily: 'Nunito',
+    paddingLeft: '9px',
+    verticalAlign: 'middle',
+  },
+  detailContainerHeader: {
+    textTransform: 'uppercase',
+    padding: '2px 8px',
+    backgroundColor: '#e6d2f2',
+    color: '#000000',
+    fontFamily: 'Nunito',
+    fontSize: '12px',
+    fontWeight: '400',
+    letterSpacing: '0.9px',
+    verticalAlign: 'middle',
+    borderRadius: '4px',
+  },
+});
+
+export default withStyles(styles, { withTheme: true })(ValueCard);
