+++ src/bento-core/GlobalSearch/index.js
@@ -1,2 +1,4 @@
 export { SearchBarGenerator } from './SearchBar/SearchBarGenerator';
 export { DEFAULT_CONFIG_SEARCHBAR } from './SearchBar/config';
+export { SearchResults } from './SearchResults/SearchResults';
+export * from './SearchResults/components/Cards';
